Fixed a compilation warning in grid_cli/wn_automation.cpp.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@64297 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/grid/grid_cli/wn_automation.cpp
@@ -83,7 +83,7 @@ bool SWorkerNodeAutomationObject::Call(const string& method,
     else if (method == "suspend") {
         bool pullback_mode = arg_array.NextBoolean(false);
         g_SuspendWorkerNode(m_WorkerNode,
-                pullback_mode, arg_array.NextInteger(0));
+                pullback_mode, (unsigned int) arg_array.NextInteger(0));
     } else if (method == "resume")
         g_ResumeWorkerNode(m_WorkerNode);
     else if (method == "shutdown")
